docs(openstack): instructions did not work.
The openstack instructions were a mix of new/old commands and
wouldn't work on a lot of openstack installs.

These documentation and changes to the provisioning script should
allow for more variation in openstack installs.

This has been tested as working against both a Havana and Juno release
of Openstack.